FTN-20 move db models
