+++ api/DHSC.FingertipsNext.Modules.Core.Service/IHealthMeasureService.cs
@@ -1,4 +1,5 @@
 ﻿using DHSC.FingertipsNext.Modules.Core.Repository;
+using DHSC.FingertipsNext.Modules.Core.Repository.Models;
 
 
 namespace DHSC.FingertipsNext.Modules.Core.Service
